Added callback functionality to Hashmap_t container.
Added similar Hashmap_DoCallback() and Hashmap_DoCallbackArg() functions
to the Hashmap_t container.

These are very similar to those for the Array_t and List_t containers,
with the exception of the "Value" pointer passed in to the callback
is actually a void*[2], containing a pointer to the Key and the Value of
the Hashmap entry respectively.

+++ src/array/array_sizing.c
@@ -79,22 +79,22 @@ int Array_Grow(Array_t *Array, size_t AdditionalCapacity) {
     return 0;
 }
 
-ssize_t Array_Length(Array_t *Array) {
+size_t Array_Length(Array_t *Array) {
 
     if (NULL == Array) {
         DEBUG_PRINTF("%s", "Error, NULL Array_t* provided.");
         return 0;
     }
 
-    return (ssize_t)Array->Length;
+    return Array->Length;
 }
 
-ssize_t Array_Capacity(Array_t *Array) {
+size_t Array_Capacity(Array_t *Array) {
 
     if (NULL == Array) {
         DEBUG_PRINTF("%s", "Error, NULL Array_t* provided.");
         return 0;
     }
 
-    return (ssize_t)Array->Capacity;
+    return Array->Capacity;
 }
